% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCFunctions.R
\name{callDynamicCloud}
\alias{callDynamicCloud}
\title{Calls the C++ dynamic_cloud.}
\usage{
callDynamicCloud(vector, nb.partitions)
}
\arguments{
\item{vector}{vector representing the data to split}

\item{nb.partitions}{number of partitions}
}
\description{
Interface to call the the C++ dynamic_cloud.
}
\author{
Raphael Couturier
}
