% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCFunctions.R
\name{callWriteTransactions}
\alias{callWriteTransactions}
\title{Calls the C++ write_transactions.}
\usage{
callWriteTransactions(data)
}
\arguments{
\item{data}{dataframe representing all the data}
}
\description{
Interface to call the the C++ write_transactions.
}
\author{
Raphael Couturier
}
