% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndAnalyzeData.R
\name{readAndAnalyzeData}
\alias{readAndAnalyzeData}
\title{Read and analyze data (preprocessing of data)}
\usage{
readAndAnalyzeData(fileName, ask.value.partitionning = FALSE)
}
\arguments{
\item{fileName}{name of the file containing data}

\item{ask.value.partitionning}{if true, user can choose the number of partition for interval variables 
(otherwise it is fixed to 3 by default)}
}
\description{
Partitions data that require that
}
\author{
Raphael Couturier
}
