% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbarGraph.R
\name{toolbarGraph}
\alias{toolbarGraph}
\title{Displays the toolbar for the implicative graph}
\usage{
toolbarGraph(frame, callPlot, updatePlot)
}
\arguments{
\item{frame}{frame to display the toolbar, the frame is created first}

\item{callPlot}{routine to compute the graph}

\item{updatePlot}{routine to update the graph}
}
\description{
Display the current parameters for the graph: thresholds, checkboxes, colors.

When the user press OK, the routine mycallPlot is called to update the graph
}
\author{
Raphael Couturier
}
