% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbarItem.R
\name{toolbarItem}
\alias{toolbarItem}
\title{Displays the toolbar so select the items to display}
\usage{
toolbarItem(list.variables, list.tcl, mycallPlot)
}
\arguments{
\item{list.variables}{list of the name of the variables}

\item{list.tcl}{list of the selected items in the tcl format}

\item{mycallPlot}{routine to call to update the graph or the tree}
}
\description{
Display the current list of selected items. F7 and F8 can be used to select or deselect all the items respectively

When the user press OK, the routine mycallPlot is called to update the graph or the tree

The toolbar is created inside another frame
}
\author{
Raphael Couturier
}
